(#355) Rename retriever_factory
