+++ NRT/nrt_collector.py
@@ -3,7 +3,7 @@ import logging
 import toml, json
 
 # Local modules
-import retriever_factory, nrtdb
+import RetrieverFactory, nrtdb
 
 dbconn = None
 ftpconn = None
@@ -28,7 +28,7 @@ for instrument_id in instruments:
   if instrument["type"] is None:
     log_instrument(instrument_id, logging.ERROR, "Configuration type not set")
   else:
-    retriever = retriever_factory.get_instance(instrument["type"],
+    retriever = RetrieverFactory.get_instance(instrument["type"],
       instrument["config"])
 
     print(retriever)
